+++ client/smart-waterloo-web/src/data/types/events.ts
@@ -16,4 +16,14 @@ const defaultEventsState: {
 	events: typeof defaultEvent[]
 } = {set: false, events:[]};
 
+interface postEventType {
+	name:string, age:string, 
+	start_day:string, start_month:string, start_year:string,
+	end_day:string, end_month:string, end_year:string,
+	category:string, description: string
+}
+
+type postEventReturn = {success:boolean, errors: string[], eventId:string}
+
 export { defaultEvent, defaultEventsState}
+export type {postEventType, postEventReturn}
